move csv2pkg command to serveradmin app
